feat(navbar): isolate navbar component
