--- src/components/Navbar/styles.ts
@@ -1,48 +1,0 @@
-import styled from 'styled-components';
-import Link from 'next/link'
-
-import styles from '@/styles/styles';
-
-interface LinkProps {
-    active?: 'true' | 'false';
-}
-
-export const Container = styled.nav`
-    display: flex;
-    align-items: center;
-    justify-content: space-between;
-    width: 100%;
-    height: 85px;
-    padding: 21px 96px;
-    background-color: ${styles.colors.white};
-    
-    font-size: ${styles.fontSizes.sm};
-    font-weight: ${styles.fontWeights.semibold};
-    text-transform: uppercase;
-`;
-
-export const NavOptions = styled.div`
-    display: flex;
-    align-items: center;
-    justify-content: space-between;
-    width: 50%;
-    gap: 25px;
-`;
-
-export const StyledLink = styled(Link)<LinkProps>`
-    color: ${({ active}) => active === 'true' ? 
-        styles.colors.primary : 'inherit'};
-`;
-
-export const StyledSelect = styled.select`
-    color: ${styles.colors.primary};
-    text-align-last: left;
-    border: none;
-    outline: none;
-    font-size: inherit;
-    font-weight: inherit;
-    font-family: inherit;
-    text-transform: inherit;
-    text-decoration: none;
-    background-color: transparent;
-`;
